added userActions
