+++ backend/pharmacy/api/views/userActions/register.py
@@ -8,30 +8,32 @@
 #
 # All rights reserved.
 
-from django.http import Http404
-from rest_framework.response import Response
-from rest_framework.decorators import permission_classes
-from rest_framework.permissions import IsAuthenticated
 
-from pharmacy.models import Medical, Medicine
-from ...serializers import MedicalSerializer, MedicineSerializer, RegisterSerializer
-
-# For customizing user claims
-from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
-from rest_framework_simplejwt.views import TokenObtainPairView
-
-# Imports for caching
-from rest_framework.views import APIView
-
-from django.utils.decorators import method_decorator
-from django.views.decorators.cache import cache_page
-from django.views.decorators.vary import vary_on_cookie, vary_on_headers
-
+from ...serializers import RegisterSerializer
 
 # Imports for registering a new user
 from rest_framework import generics
 from rest_framework.permissions import AllowAny
 from django.contrib.auth.models import User
 
-# Imports used in search functionality
-from rest_framework import filters
+class Register(generics.CreateAPIView):
+    """
+    CreateAPIView is a generic class-based view that allows you to handle POST requests.
+
+    This class is used to register a new user.
+    """
+
+    queryset = User.objects.all()
+    """
+    Queryset is used to get all the users
+    """
+
+    permission_classes = (AllowAny,)
+    """
+    Allow only unauthenticated users to access this view
+    """
+
+    serializer_class = RegisterSerializer
+    """
+    Now registering the serializer we created in serializers.py
+    """
